feat: tweak statusline and show cmds
# Status Line

There was a problem I had with the statusline, when I open a buffer,
it's good, but if I edit it, it goes blank. These changes now prevent
this and also show the filename if the buftype is "nofile" as is the
case of nvim-tree. 

# Show CMDs

Also show again cmds because that's preventing me
from seeing the registers I type (i.e. "2) so even though it's a bit
annoying, it's more helpful than no cmd at all.